requirements.txt
websockets>=12.0
aiohttp>=3.9
python-dotenv>=1.0
py-clob-client>=0.18
web3>=6.0
eth-account>=0.10
python-telegram-bot>=20.0
pandas>=2.0
pytest>=7.4
pytest-asyncio>=0.23
rich>=13.0
